helper-cli: Move all commands to the "commands" package
And allow wilcard imports from the "commands" package for convenience.
